requirements.txt
load_dotenv
os
pandas
streamlit
supabase

dash
plotly
gunicorn
